requirements.txt
asgiref==3.7.2
asyncio==3.4.3
backports.zoneinfo==0.2.1
Django==4.2.8
sqlparse==0.4.4
typing-extensions==4.9.0
websocket-client==1.7.0
websockets==12.0
